+++ Plugin/Selva.Drawing/SvgSurfaceData.cs
@@ -1,15 +1,14 @@
 using System.Collections.Generic;
-using Rhino.Geometry;
 
-namespace Selva.GH.Features.Drawing.Lib;
+namespace Selva.Drawing;
 
 public class SvgSurfaceData
 {
     public string OuterPathData { get; set; }
-    public List<string> HolePathData { get; set; } = new();
-    public BoundingBox Bounds { get; set; }
+    public List<string> HolePathData { get; set; } = new List<string>();
+    public SvgBounds Bounds { get; set; }
     public PathStyleData Style { get; set; }
-    public Dictionary<string, string> Metadata { get; set; } = new();
+    public Dictionary<string, string> Metadata { get; set; } = new Dictionary<string, string>();
     public string Id { get; set; }
     public string CssClass { get; set; }