+++ Plugin/Selva.GH/Features/Drawing/Components/GH_CreateSvgSurface.cs
@@ -1,9 +1,9 @@
 using System;
-using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
 using Grasshopper.Kernel;
 using Rhino.Geometry;
+using Selva.Drawing;
 using Selva.GH.Features.Drawing.Lib;
 
 namespace Selva.GH.Features.Drawing.Components;
@@ -19,24 +19,14 @@ public class GH_CreateSvgSurface : GH_Component
 
     protected override Bitmap Icon => null;
     public override GH_Exposure Exposure => GH_Exposure.primary;
-    public override Guid ComponentGuid => new Guid("1BFCD1BD-DA7F-4ED2-AE8D-34EA2C347B85");
+    public override Guid ComponentGuid => new Guid("5D7E9A2B-3C4F-4B8E-A1D2-6F8B9C0E1F23");
 
     protected override void RegisterInputParams(GH_InputParamManager pManager)
     {
         pManager.AddBrepParameter("Brep", "B", "Input Brep", GH_ParamAccess.item);
         pManager.AddGenericParameter("Style", "S", "Path style (use Path Style component)", GH_ParamAccess.item);
-        pManager.AddTextParameter("ID", "ID", "Element id", GH_ParamAccess.item, "");
-        pManager.AddTextParameter("CSS Class", "Cls", "CSS class", GH_ParamAccess.item, "");
-        pManager.AddTextParameter("Metadata Keys", "MK", "Metadata keys", GH_ParamAccess.list);
-        pManager.AddTextParameter("Metadata Values", "MV", "Metadata values", GH_ParamAccess.list);
-        pManager.AddNumberParameter("Tolerance", "T", "Tolerance for joining edges", GH_ParamAccess.item, 0.01);
 
         pManager[1].Optional = true;
-        pManager[2].Optional = true;
-        pManager[3].Optional = true;
-        pManager[4].Optional = true;
-        pManager[5].Optional = true;
-        pManager[6].Optional = true;
     }
 
     protected override void RegisterOutputParams(GH_OutputParamManager pManager)
@@ -48,19 +38,10 @@ public class GH_CreateSvgSurface : GH_Component
     {
         Brep brep = null;
         PathStyleData style = null;
-        var id = "";
-        var cssClass = "";
-        var metaKeys = new List<string>();
-        var metaValues = new List<string>();
         var tolerance = 0.01;
 
         if (!DA.GetData(0, ref brep) || brep == null) return;
         DA.GetData(1, ref style);
-        DA.GetData(2, ref id);
-        DA.GetData(3, ref cssClass);
-        DA.GetDataList(4, metaKeys);
-        DA.GetDataList(5, metaValues);
-        DA.GetData(6, ref tolerance);
 
         try
         {
@@ -82,11 +63,6 @@ public class GH_CreateSvgSurface : GH_Component
                 .Where(s => !string.IsNullOrEmpty(s))
                 .ToList();
 
-            var metadata = new Dictionary<string, string>();
-            for (var i = 0; i < Math.Min(metaKeys.Count, metaValues.Count); i++)
-                if (!string.IsNullOrEmpty(metaKeys[i]))
-                    metadata[metaKeys[i]] = metaValues[i] ?? "";
-
             var fillStyle = style ?? new PathStyleData
             {
                 FillColor = Color.LightGray,
@@ -96,15 +72,13 @@ public class GH_CreateSvgSurface : GH_Component
                 StrokeWidth = 1f
             };
 
+            var bb = brep.GetBoundingBox(true);
             var data = new SvgSurfaceData
             {
                 OuterPathData = outerPath,
                 HolePathData = holePaths,
-                Bounds = brep.GetBoundingBox(true),
-                Style = fillStyle,
-                Id = id,
-                CssClass = cssClass,
-                Metadata = metadata
+                Bounds = new SvgBounds(bb.Min.X, bb.Min.Y, bb.Max.X, bb.Max.Y),
+                Style = fillStyle
             };
 
             DA.SetData(0, data);